// src/components/ExImage/types.ts
export type ExImagePropsType = {
    directory?: string
    isPublicDir?: boolean
    cover?: boolean
    contain?: boolean

    size?: string | number
    radius?: string | number

    width?: string | number
    height?: string | number
    minWidth?: string | number
    minHeight?: string | number
    maxWidth?: string | number
    maxHeight?: string | number
}